♻️ refactor: split cache.ts into config/client/helpers modules #15
