+++ src/lib/cache/helpers.ts
@@ -1,0 +1,40 @@
+import { logger } from '../logger';
+import { cache } from './client';
+
+/**
+ * キャッシュキーを生成するヘルパー関数
+ */
+export function generateCacheKey(prefix: string, params: Record<string, string | number>): string {
+  const sortedParams = Object.keys(params)
+    .sort()
+    .map((key) => `${key}:${params[key]}`)
+    .join(':');
+  return `${prefix}:${sortedParams}`;
+}
+
+/**
+ * キャッシュを通してデータを取得するヘルパー関数
+ * キャッシュミス時はfetcherを実行してキャッシュに保存
+ */
+export async function withCache<T>(
+  key: string,
+  ttlSeconds: number,
+  fetcher: () => Promise<T>,
+): Promise<{ data: T; fromCache: boolean }> {
+  // キャッシュから取得を試みる
+  const cached = await cache.get<T>(key);
+
+  if (cached !== null) {
+    return { data: cached, fromCache: true };
+  }
+
+  // fetcherでデータを取得
+  const data = await fetcher();
+
+  // キャッシュに保存（非同期、エラーは無視）
+  cache.set(key, data, ttlSeconds).catch((error) => {
+    logger.error({ key, error }, 'Failed to cache data');
+  });
+
+  return { data, fromCache: false };
+}
